requirements.txt
pyserial==3.4
python-dateutil==2.8.0